+++ tests/cache/test_cache.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from unittest.mock import MagicMock
+from unittest.mock import AsyncMock
 
 from pytest import fixture, mark
 
@@ -8,15 +8,16 @@ from limberframework.cache.cache import Cache
 
 @fixture
 def mock_store():
-    return MagicMock()
+    return AsyncMock()
 
 
-def test_load(mock_store):
+@mark.asyncio
+async def test_load(mock_store):
     data = {"data": "test", "expires_at": datetime.now()}
-    mock_store.__getitem__.return_value = data
+    mock_store.get.return_value = data
 
     cache = Cache(mock_store)
-    cache.load("test_cache")
+    await cache.load("test_cache")
 
     assert cache._key == "test_cache"
     assert cache.value == data["data"]
@@ -32,7 +33,8 @@ def test_load(mock_store):
         ("test", None, False),
     ],
 )
-def test_update(value, expires_at, updated, mock_store):
+@mark.asyncio
+async def test_update(value, expires_at, updated, mock_store):
     mock_store.put.return_value = True
     key = "test_key"
 
@@ -41,6 +43,6 @@ def test_update(value, expires_at, updated, mock_store):
     cache.value = value
     cache.expires_at = expires_at
 
-    response = cache.update()
+    response = await cache.update()
 
     assert response == updated