+++ tests/routing/test_rate_limiter.py
@@ -1,11 +1,11 @@
 from datetime import datetime, timedelta
 from math import ceil
-from unittest.mock import patch
+from unittest.mock import AsyncMock, patch
 
 from pytest import mark, raises
 
 from limberframework.routing.exceptions import TooManyRequestsException
-from limberframework.routing.rate_limiter import RateLimiter
+from limberframework.routing.rate_limiter import RateLimiter, make_rate_limiter
 
 
 @patch("limberframework.routing.rate_limiter.Cache")
@@ -19,26 +19,28 @@ def test_get_hits(mock_cache, value, hits):
     assert response == hits
 
 
-@patch("limberframework.routing.rate_limiter.Cache")
-def test_set_hits(mock_cache):
+@patch("limberframework.routing.rate_limiter.Cache", new_callable=AsyncMock)
+@mark.asyncio
+async def test_set_hits(mock_cache):
     hits = 10
     date = datetime.now()
     mock_cache.expires_at = date
 
-    rate_limiter = RateLimiter(mock_cache, "test", 60, 60)
-    rate_limiter.set_hits(hits)
+    rate_limiter = await make_rate_limiter(mock_cache, "test", 60, 60)
+    await rate_limiter.set_hits(hits)
 
     assert mock_cache.value == str(hits)
     mock_cache.update.assert_called_once()
 
 
-@patch("limberframework.routing.rate_limiter.Cache")
-def test_set_hits_no_expiry(mock_cache):
+@patch("limberframework.routing.rate_limiter.Cache", new_callable=AsyncMock)
+@mark.asyncio
+async def test_set_hits_no_expiry(mock_cache):
     hits = 10
     mock_cache.expires_at = None
 
-    rate_limiter = RateLimiter(mock_cache, "test", 60, 60)
-    rate_limiter.set_hits(hits)
+    rate_limiter = await make_rate_limiter(mock_cache, "test", 60, 60)
+    await rate_limiter.set_hits(hits)
 
     assert mock_cache.value == str(hits)
     assert mock_cache.expires_at is not None
@@ -86,23 +88,25 @@ def test_available_in_no_expiry(mock_cache, mock_date_time):
     assert response == ceil((expiry - now).total_seconds())
 
 
-@patch("limberframework.routing.rate_limiter.Cache")
-def test_hit_exception(mock_cache):
+@patch("limberframework.routing.rate_limiter.Cache", new_callable=AsyncMock)
+@mark.asyncio
+async def test_hit_exception(mock_cache):
     hits = 60
     mock_cache.value = hits
 
-    rate_limiter = RateLimiter(mock_cache, "test", hits, 60)
+    rate_limiter = await make_rate_limiter(mock_cache, "test", hits, 60)
 
     with raises(TooManyRequestsException):
-        rate_limiter.hit()
+        await rate_limiter.hit()
 
 
-@patch("limberframework.routing.rate_limiter.Cache")
-def test_hit(mock_cache):
+@patch("limberframework.routing.rate_limiter.Cache", new_callable=AsyncMock)
+@mark.asyncio
+async def test_hit(mock_cache):
     hits = 40
     mock_cache.value = hits
 
-    rate_limiter = RateLimiter(mock_cache, "test", 60, 60)
-    rate_limiter.hit()
+    rate_limiter = await make_rate_limiter(mock_cache, "test", 60, 60)
+    await rate_limiter.hit()
 
     assert mock_cache.value == str(hits + 1)